requirements.txt
matplotlib>=3.5.1
numpy>=1.21.5
scipy>=1.8.0
pyade-python>=1.1